refactored folder placement of wiki accrodian and widgets
